==> 2.Table1.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt
import wrds
import matplotlib.pyplot as plt
from dateutil.relativedelta import *
from pandas.tseries.offsets import *
import warnings
warnings.filterwarnings('ignore')
import pingouin as pg
import statsmodels.api as sm

data = pd.read_pickle('data99.pkl')
data_use = data.copy()

factors = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']
# data_use = data_use.drop(columns=[col for col in factors if (data_use[col] == 0).all()])


#########################
### Table 1 Panel B #####
#########################

factors = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']

for factor in factors:
    data_use[f'{factor}_me'] = data_use[factor] * data_use['me']
grouped = data_use.groupby('jdate')[[f'{factor}_me' for factor in factors]].sum()
grouped['total_me'] = data_use.groupby('jdate')['me'].sum()
for factor in factors:
    grouped[factor] = grouped[f'{factor}_me'] / grouped['total_me']
avg_market_cap_percentage = grouped[factors].mean() * 100
avg_market_cap_percentage = avg_market_cap_percentage.round(1)

table1_panelB = pd.DataFrame([avg_market_cap_percentage.values], columns=factors)
table1_panelB.index = ['Average percent of MC']
table1_panelB

## 13.9 11.6 2.3 51.9 48.4 3.5 ##

#########################
### Table 1 Panel A #####
#########################
data_use = data.copy()

before = data_use[(data_use['jdate']>='1963-07-31') & (data_use['jdate']<='1991-06-30')]
after = data_use[(data_use['jdate']>'1991-06-30') & (data_use['jdate']<='2019-06-30')]
whole = data_use.copy()

def wavg(group, avg_name, weight_name):
    d = group[avg_name]
    w = group[weight_name]
    try:
        return (d * w).sum() / w.sum()
    except ZeroDivisionError:
        return np.nan

def calculate_portfolio_returns(data, portfolios,var,weight):
    result = pd.DataFrame()

    for portfolio in portfolios:
        subset = data[data[portfolio] == 1]
        result[portfolio] = subset.groupby('jdate').apply(wavg, var, weight)
    result['all_vwret'] = data.groupby('jdate').apply(wavg, var, weight)

    for portfolio in portfolios:
        result[portfolio + '_excess'] = result[portfolio] - result['all_vwret']

    portfolio_return = result.reset_index()
    return portfolio_return

portfolios = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']

portfolio_return = calculate_portfolio_returns(whole,portfolios,'ret_ahead','me')

portfolio_return.to_pickle('portfolio_return.pkl')

portfolio_return00 = portfolio_return.copy()
portfolio_return00['post_dummy'] = (portfolio_return00['jdate'] <= '1991-06-30').astype(int)
subset = portfolio_return00.copy()
subset00 = portfolio_return00[portfolio_return00['post_dummy']==1].reset_index(drop=True)
subset11 = portfolio_return00[portfolio_return00['post_dummy']==0].reset_index(drop=True)
subset99 = subset00 - subset11

def summary(data, portfolio):
    portfolio_name = portfolio + '_excess'
    _data00 = data[portfolio_name]
    _data00 = _data00

    mean = _data00.mean()  
    var = _data00.var(ddof=1)  
    sd = var ** 0.5  
    n = len(_data00)  
    se = (var / n) ** 0.5 
    t_stat_diff = mean / se 

    result = pd.DataFrame(
        [
            mean ,  
            sd ,   
            t_stat_diff  
        ],
        index=['Average Excess Return', 'Standard Deviation', 't-Statistic'],
        columns=[portfolio]
    )

    return result

result00 = pd.concat(
    [summary(subset, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

result11 = pd.concat(
    [summary(subset00, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

result22 = pd.concat(
    [summary(subset11, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

result33 = pd.concat(
    [summary(subset99, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

table1_panelA = pd.concat([result00,result11,result22,result33], axis=0)
table1_panelA

result00 = pd.concat(
    [summary(subset, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

result11 = pd.concat(
    [summary(subset00, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

result22 = pd.concat(
    [summary(subset11, portfolio=portfolio) for portfolio in portfolios], 
    axis=1
).round(2)

portfolio_return00 = portfolio_return.copy()
portfolio_return00['post_dummy'] = (portfolio_return00['jdate'] <= '1991-06-30').astype(int)

result33 = pd.DataFrame(columns=['MV', 'BV', 'SV', 'MG', 'BG', 'SG'])

portfolios = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']
for portfolio in portfolios:
    y = portfolio_return00[portfolio + '_excess']  
    X = sm.add_constant(portfolio_return00['post_dummy']) 
    model = sm.OLS(y, X).fit()

    coefficient = model.params['post_dummy']
    std_err = model.bse['post_dummy']
    std_resid = model.mse_resid ** 0.5
    t_stat = model.tvalues['post_dummy']
    result33[portfolio] = [coefficient , std_resid , t_stat]
result33.index = ['Average Excess Return', 'Standard Deviation', 't-Statistic']

table1_panelA_v2 = pd.concat([result00,result11,result22,result33], axis=0).round(2)
table1_panelA_v2

portfolio_return11 = portfolio_return.copy()

#########################
### Table 1 Panel C #####
#########################

variables = ['all_vwret', 'MV', 'BV', 'SV', 'MG', 'BG', 'SG']
subset = portfolio_return11[variables] 
corr_matrix = subset.corr() 
corr_matrix00 = pd.DataFrame(corr_matrix.iloc[0, 1:].round(2)).T 

variables = ['MV_excess', 'BV_excess', 'SV_excess', 'MG_excess', 'BG_excess', 'SG_excess']
subset = portfolio_return11[variables]
corr_matrix11 = subset.corr().round(2)  
corr_matrix11.columns = corr_matrix11.columns.str.replace('_excess', '')  
corr_matrix11.index = corr_matrix11.index.str.replace('_excess', '') 

corr_matrix = pd.concat([corr_matrix00, corr_matrix11])
corr_matrix.index.values[0] = 'Cor(R, RM)' 
table1_panelC = corr_matrix

#########################
###  Hotelling T^2  #####
#########################
X = subset00[variables]
Y = subset11[variables]
print(pg.multivariate_ttest(X, Y,paired=False))

print(table1_panelA)
print(pg.multivariate_ttest(X, Y,paired=False))
print(table1_panelB)
print(table1_panelC)

table1_panelA.to_excel("/Users/fuhk/Downloads/table_panelA.xlsx")
