==> 4.Table3.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt
import wrds
import matplotlib.pyplot as plt
from dateutil.relativedelta import *
from pandas.tseries.offsets import *
from scipy import stats
import janitor 
import warnings
warnings.filterwarnings('ignore')
import seaborn as sns

import statsmodels.api as sm

data = pd.read_pickle("data99.pkl")
beme = pd.read_pickle('portfolio_beme.pkl')
returns = pd.read_pickle('portfolio_return.pkl')

list = ['MV_excess', 'BV_excess', 'SV_excess', 'MG_excess', 'BG_excess', 'SG_excess']
beme = beme.rename(columns={col: col.replace('_excess', '_BM') for col in list})
beme00 = beme[['jdate','MV_BM','BV_BM','SV_BM','MG_BM','BG_BM','SG_BM']]
returns = returns.rename(columns={col: col.replace('_excess', '_Return') for col in list})
returns00 = returns[['jdate','MV_Return','BV_Return','SV_Return','MG_Return','BG_Return','SG_Return']]

sample = pd.merge(returns00,beme00,on='jdate',how='inner')
sample_whole = sample[(sample['jdate']>='1963-07-31') & (sample['jdate']<='2019-06-30')]
sample_before = sample[(sample['jdate']>='1963-07-31') & (sample['jdate']<='1991-06-30')]
sample_after = sample[(sample['jdate']>'1991-06-30') & (sample['jdate']<='2019-06-30')]

portfolios = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']
results = []

def portfolio_regressions(data, portfolios):
    results = []
    for portfolio in portfolios:
        ret_col = f"{portfolio}_Return" 
        bm_col = f"{portfolio}_BM"     
        
        X = sm.add_constant(data[bm_col])  
        y = data[ret_col]
        model = sm.OLS(y, X).fit()       

        results.append({
            'portfolio': portfolio,
            'a': model.params['const'],                # 截距
            'b': model.params[bm_col],                # 斜率
            'StdErr(a)': model.bse['const'],          # 截距标准误差
            'StdErr(b)': model.bse[bm_col],           # 斜率标准误差
            't(a)': model.tvalues['const'],           # 截距 t 值
            't(b)': model.tvalues[bm_col],            # 斜率 t 值
            'R2': model.rsquared,                     # R 方
            'RSE': (model.ssr / model.df_resid) ** 0.5  # 残差标准误差
        })
    return pd.DataFrame(results)

    

portfolios = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']
table3_panelA = portfolio_regressions(sample_whole, portfolios)
table3_panelA

sample_whole = sample.copy()
sample_whole['dummy'] = (sample_whole['jdate'] > '1991-06-30').astype(int)

portfolios = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']
results = []

def portfolio_regressions_with_dummy(data, portfolios, dummy_col):
    results = []
    cov_matrices = []  
    
    for portfolio in portfolios:
        ret_col = f"{portfolio}_Return"
        bm_col = f"{portfolio}_BM"
        
        data['const_dummy'] = data[dummy_col]
        data['bm_dummy'] = data[dummy_col] * data[bm_col]

        X = sm.add_constant(data[[ 'const_dummy', 'bm_dummy',bm_col]])
        y = data[ret_col]
        model = sm.OLS(y, X).fit()
        
        cov_matrix = model.cov_params() 
        cov_matrices.append(cov_matrix) 
        
        results.append({
            'portfolio': portfolio,
            'a': model.params['const'], 
            'da': model.params['const_dummy'],                
            'b': model.params[bm_col],                
            'db': model.params['bm_dummy'],      
            'StdErr(a)': model.bse['const'], 
            'StdErr(da)': model.bse['const_dummy'],        
            'StdErr(b)': model.bse[bm_col],          
            'StdErr(db)': model.bse['bm_dummy'],    
            't(a)': model.tvalues['const'],  
            't(da)': model.tvalues['const_dummy'],         
            't(b)': model.tvalues[bm_col],          
            't(db)': model.tvalues['bm_dummy'],     
            'R2': model.rsquared,                     
            'RSE': (model.ssr / model.df_resid) ** 0.5  
        })
    
    results_df = pd.DataFrame(results)
    
    return results_df, cov_matrices

portfolios = ['MV', 'BV', 'SV', 'MG', 'BG', 'SG']
table3_panelB,cov= portfolio_regressions_with_dummy(sample_whole, portfolios,dummy_col='dummy')
table3_panelB

X1_mean = sample_whole[sample_whole['dummy'] == 0][['MV_BM', 'BV_BM', 'SV_BM', 'MG_BM', 'BG_BM', 'SG_BM']].mean()
X1_mean = X1_mean.reset_index().rename(columns={'index': 'portfolio', 0: 'X1_mean'})
X1_mean['portfolio'] = X1_mean['portfolio'].str.replace('_BM', '')

X2_mean = sample_whole[sample_whole['dummy'] == 1][['MV_BM', 'BV_BM', 'SV_BM', 'MG_BM', 'BG_BM', 'SG_BM']].mean()
X2_mean = X2_mean.reset_index().rename(columns={'index': 'portfolio', 0: 'X2_mean'})
X2_mean['portfolio'] = X2_mean['portfolio'].str.replace('_BM', '')

table00 = pd.merge(table3_panelA,X1_mean,on='portfolio',how='inner')
table11 = pd.merge(table00,X2_mean,on='portfolio',how='inner')
table11['Y11'] = table11['a'] + table11['X1_mean']*table11['b'] 
table11['Y12'] = table11['a'] + table11['X2_mean']*table11['b']
table11['Y11-Y12'] = table11['Y11'] - table11['Y12']
table11['se(Y11-Y12)'] = np.abs(((table11['X1_mean'] - table11['X2_mean'])*table11['StdErr(b)']))
table11['t(Y11-Y12)'] = table11['StdErr(b)']*np.sign(table11['X1_mean'] - table11['X2_mean'])
table11

table22 = pd.merge(table3_panelB,X1_mean, on='portfolio',how='inner')
table33 = pd.merge(table22,X2_mean,on='portfolio',how='inner')
table33['Y21'] = table33['a'] + table33['b']*table33['X1_mean']
table33['Y22'] = table33['a'] + table33['da'] + table33['b']*table33['X2_mean'] + table33['db']*table33['X2_mean']
table33['Y21-Y22'] = table33['Y21'] - table33['Y22']

cov_matrices = []

for cov_matrix in cov:
    cov_matrix = cov_matrix.drop(index='const', columns='const')
    cov_matrices.append(cov_matrix)

def calculate_se(row, cov_matrix):

    V = np.array([1, row['X2_mean'],row['X2_mean'] - row['X1_mean']]) 

    se = np.sqrt(V.T @ cov_matrix @ V)
    return se

table33['se(Y21-Y22)'] = table33.apply(
    lambda row: calculate_se(row, cov_matrices[table33.index.get_loc(row.name)]), axis=1
)

table33['t(Y21-Y22)'] = table33['Y21-Y22'] / table33['se(Y21-Y22)']

table3_panelC = pd.merge(table11[['portfolio','X1_mean','X2_mean','Y11',"Y12",'Y11-Y12','se(Y11-Y12)','t(Y11-Y12)']],
                        table33[['portfolio','Y21',"Y22",'Y21-Y22','se(Y21-Y22)','t(Y21-Y22)']], on='portfolio', how='outer')

print(table3_panelA)
print(table3_panelB)
print(table3_panelC)
